feat(conf) implement plugins option and deprecate custom_plugins
Problem:
All plugins bundled with Kong are loaded
by default. There exists no mechanism
to disable any of the bundled plugins.

Solution:
Introduce a config option which
controls which of the bundled plugins
are loaded into Kong.

Implementation:
Introduce `plugins` config option and
deprecate `custom_plugins` option. Kong will
load only plugins which are specified in `plugins`
config options. This also is inline in the goal
of simplifying Kong's configurations.

Plugins specified in `plugins` and `custom_plugins`
are merged together to avoid a break change but
`custom_plugins` will be removed in future.

The new `plugins` options can take in a sugar
value `bundled` which loads all the plugins bundled with Kong.
This accomplishes  good usability and flexibility.